Changed semantics of "default" tag
The "default" tag now supplies a real default value for a flag
value. Even though default values are best set in the struct
directly, sometimes it's useful to have flags be represented
by callback functions, which can't really have default values
(in the sense of flags). For those flags, the default:'val" tag
now allows you to set a default value (it will simply invoke the flag with the default value).

The previous meaning of default is now provided by the optional-value:"val" tag.