Support Windows-style options.
This commit adds support for front slash to be used for options on Windows.

Front slashes are the standard style for Windows CLI programs.  Also,
Windows doesn't typically differentiate between short and long options
because a front slash works for both.  Finally, Windows CLI applications
typically support /? for help.

So given the above, this commit makes the following modifications only
when running on Windows:

- Adds support for Windows-style options with short names (/v)
- Adds support for Windows-style options with long names (/verbose)
- Adds an additional Windows-style help entry (/?)
- Modifies the help (usage) display to show Windows-style options
  (example: /v for short names, /verbose for long names)
- Note that in all cases the unix style -, and -- still work as well
  on Windows for consistency among operating systems.

Finally, the documentation has been updated to reflect the new
capabilities.