Improve support for Windows-style options.
This commit expands upon the previous commit to add support for
Windows-style options.

At the request of the go-flags author, the code has been modified to
use build constraints for windows-specific functionality instead of
selecting at runtime.

As noted by @jessvdk, Windows typically uses a colon for the option name
and argument delimiter.  This code adds support for both styles, but
doesn't allow the two to be mixed.  That is to say /foo:bar and --foo=bar
are acceptable, but /foo=bar and --foo:bar are not.