add support for requiring `rest` positional args
By default the slice in the end of positional args is not required, but often that
sort of slice is used for example in cases where multiple file names are passed
and at least one is required. To support this use case, add support for specifying
`required` tag for the last slice of positional args and check that when verifying
command arguments.